+++ .nycrc
@@ -4,5 +4,6 @@
   "functions": 100,
   "statements": 100,
   "check-coverage": true,
-  "reporter": ["lcov", "text"]
+  "reporter": ["lcov", "text"],
+  "exclude": ["./sheetsSetup.js"]
 }
