+++ lib/apiForSheets.js
@@ -13,16 +13,11 @@
 // limitations under the License.
 
 // Google Sheets API v4
-const readline = require('readline');
 const { google } = require('googleapis');
 const { OAuth2Client } = require('google-auth-library');
 const claLabel = "PR: don't merge - NEEDS CLA";
-const claLabelArray = [claLabel];
 var spreadsheetId = process.env.SPREADSHEET_ID;
 
-// If modifying these scopes, delete your previously saved credentials
-// at ~/.credentials/sheets.googleapis.com-nodejs-quickstart.json
-var SCOPES = ['https://www.googleapis.com/auth/spreadsheets.readonly'];
 var clientSecret = process.env.CLIENT_SECRET;
 
 /**
@@ -168,35 +163,3 @@ module.exports.checkClaStatus = async function (contextData) {
   // Call the sheets API with the authorized client.
   module.exports.checkClaSheet(oauthClient);
 };
-
-
-/**
- * Get and store new token after prompting for user authorization, and then
- * execute the given callback with the authorized OAuth2 client.
- * @param {google.auth.OAuth2} oAuth2Client The OAuth2 client to get token for.
- * @param {getEventsCallback} callback The callback for the authorized client.
- */
-module.exports.getNewToken = (oAuth2Client, callback) => {
-  const authUrl = oAuth2Client.generateAuthUrl({
-    access_type: 'offline',
-    scope: SCOPES,
-  });
-  console.log('Authorize this app by visiting this url:', authUrl);
-  const rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout,
-  });
-  rl.question('Enter the code from that page here: ', (code) => {
-    rl.close();
-    oAuth2Client.getToken(code, (err, token) => {
-      if (err) {
-        return console.error(
-          'Error while trying to retrieve access token', err
-        );
-      }
-      oAuth2Client.setCredentials(token);
-      callback(token);
-    });
-  });
-};
-
